+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 02-01-2022 -- v0.0.1-BETA
 
  - The first version of RKimer is released upon the world.
+
+## 07-02-2022 -- v0.0.3-BETA
+
+ - Solves are now stored with their respective scramble.
+ - Added a solve data viewer with a graph of solve time, average of 3 and personal best